Add new DBOS wait timer call

+++ src/ee/voltdbjni.cpp
@@ -306,9 +306,9 @@ SHAREDLIB_JNIEXPORT jint JNICALL Java_org_voltdb_jni_ExecutionEngine_DBOSPVWait(
 
 SHAREDLIB_JNIEXPORT jint JNICALL Java_org_voltdb_jni_ExecutionEngine_DBOSPVWaitTimer(JNIEnv *env, jclass obj, jint hypervisor_fd, jint dual_qemu_pid, jint dual_qemu_lapic_id, jint wakeup_delay_us) {
     uint64_t word = 
-              ((dual_qemu_pid & 0x00000000ffffffff) << 32L) 
-            | ((wakeup_delay_us & 0x0000000000ffffff) << 8L) 
-            | ((dual_qemu_lapic_id & 0x00000000000000ff) << 0L);
+              (((uint64_t) dual_qemu_pid &      0x00000000ffffffff) << 32UL) 
+            | (((uint64_t) wakeup_delay_us &    0x0000000000ffffff) << 8UL) 
+            | (((uint64_t) dual_qemu_lapic_id & 0x00000000000000ff) << 0UL);
     return reinterpret_cast<jint>(ioctl(hypervisor_fd, DBOS_PV_WAIT_TIMER, &word));
 }
